tests(introspection): update snapshots
