+++ src/packages/introspection/src/__tests__/init.test.ts
@@ -13,21 +13,7 @@ describe('init', () => {
     const init = Init.new()
     const result = stripAnsi(await init.parse([]))
 
-    expect(result).toMatchInlineSnapshot(`
-          "
-          ✔ Your Prisma schema was created at prisma/schema.prisma.
-            You can now open it in your favorite editor.
-
-          Next steps:
-          1. Set the DATABASE_URL in the .env file to point to your existing database. If your database has no tables yet, read https://pris.ly/d/getting-started.
-          2. Set the provider of the datasource block in schema.prisma to match your database: postgresql, mysql or sqlite.
-          3. Run prisma introspect to turn your database schema into a Prisma data model.
-          4. Run prisma generate to install Prisma Client. You can then start querying your database.
-
-          More information in our documentation:
-          https://pris.ly/d/getting-started
-              "
-      `)
+    expect(result).toMatchSnapshot()
 
     const schema = fs.readFileSync(
       join(tmpDir, 'prisma', 'schema.prisma'),
@@ -50,19 +36,7 @@ describe('init', () => {
     const result = stripAnsi(
       await init.parse(['--url', process.env.TEST_POSTGRES_URI!]),
     )
-    expect(result).toMatchInlineSnapshot(`
-      "
-      ✔ Your Prisma schema was created at prisma/schema.prisma.
-        You can now open it in your favorite editor.
-
-      Next steps:
-      1. Run prisma introspect to turn your database schema into a Prisma data model.
-      2. Run prisma generate to install Prisma Client. You can then start querying your database.
-
-      More information in our documentation:
-      https://pris.ly/d/getting-started
-          "
-    `)
+    expect(result).toMatchSnapshot()
 
     const schema = fs.readFileSync(
       join(tmpDir, 'prisma', 'schema.prisma'),
